Added sample to show how to scale a cluster
